feat: Change name from `NotificationsModal` to `NotificationBoxPopup`
